Fix install_dependencies on archlinux, update readme accordingly

+++ extlibs/install_dependencies.sh
@@ -2,6 +2,37 @@
 
 set -e
 
-# Ogre3d
-##############################################################################
-sudo apt install xaw3dg xorg-dev libz-dev
+echo "This script is about to install a some packages..."
+echo "You might want to look over what it does before contining"
+read -p "Press enter to continue, or Ctrl-C to exit"
+
+if [ `command -v apt` ]
+then
+		echo "Installing packages with apt..."
+		# Build utils
+		sudo apt install cmake
+		# Ogre3d
+		sudo apt install xaw3dg xorg-dev libz-dev
+elif [ `command -v yum` ]
+then
+		echo "Broken on Centos at the moment..."
+		# Seems to be some error where cmake cant work out if pthreads is installed
+		# Getting the exact same error as on:
+		# https://cmake.org/pipermail/cmake/2015-December/062283.html
+
+		#echo "Installing packages with yum..."
+		# Build utils
+		#sudo yum install cmake gcc-c++
+		# Ogre
+		#sudo yum install mesa-libGL-devel mesa-libGLU-devel libXaw-devel
+
+elif [ `command -v pacman` ] # Keep this last since it will launch pacman game on ubuntu...
+then
+		echo "Installing packages with pacman..."
+		pacman -S cmake base-devel
+		# Ogre
+		pacman -S mesa glu
+else
+		echo "Failed to find a suitable package manager with which to install dependencies"
+		exit 1
+fi
